Renaming/restructuring and code cleanup
